backend/requirements.txt
Flask
Flask-SocketIO
Flask-Cors
paho-mqtt
gunicorn
python-dotenv
